add simple middlewares to invocation and functions, implement recursi…

+++ src/api/functions/functions.js
@@ -1,15 +1,31 @@
 'use strict'
+const AccountModel = require('../../shared/model/account')
+const UserModel = require('../../shared/model/user')
+const UserAccountModel = require('../../shared/model/userAccount')
+const passport = require('./../passport/passport')
+const errors = require('../../shared/helper/errors')
 const FunctionModel = require('../../shared/model/function')
 const InvocationModel = require('../../shared/model/invocation')
 const _ = require('lodash')
-const passport = require('./../passport/passport')
 const responses = require('../../shared/helper/responses')
 const faker = require('faker')
 const BigInteger = require('big-integer')
 
 module.exports.list = (event, context, callback) => {
-  return passport.checkAuth(event.headers.Authorization)
-    .then((user) => {
+  const token = event.headers.Authorization
+  const accountId = event.pathParameters.id
+
+  return Promise.all([
+    passport.checkAuth(token).then(decoded => UserModel.isActiveOrThrow(decoded)),
+    AccountModel.getActiveByIdrOrThrow(accountId),
+    UserAccountModel.getUsersByAccount(accountId)]
+  )
+    .then(([id, account, accountUsers]) => {
+      let accountUsersIds = accountUsers.map(user => user._user)
+      if (!(accountUsersIds.includes(id))) {
+        throw errors.forbidden()
+      }
+
       return FunctionModel.getAll({
         IndexName: 'AccountIdIndex',
         KeyConditionExpression: '#account = :account',
@@ -28,9 +44,25 @@ module.exports.list = (event, context, callback) => {
 }
 
 module.exports.get = (event, context, callback) => {
-  return passport.checkAuth(event.headers.Authorization)
-    .then((user) => {
-      return FunctionModel.getById(event.pathParameters.functionId)
+  const token = event.headers.Authorization
+  const accountId = event.pathParameters.id
+  const functionId = event.pathParameters.functionId
+
+  return Promise.all([
+    passport.checkAuth(token).then(decoded => UserModel.isActiveOrThrow(decoded)),
+    AccountModel.getActiveByIdrOrThrow(accountId),
+    UserAccountModel.getUsersByAccount(accountId)]
+  )
+    .then(([id, account, accountUsers]) => {
+      let accountUsersIds = accountUsers.map(user => user._user)
+      if (!(accountUsersIds.includes(id))) {
+        throw errors.forbidden()
+      }
+      return FunctionModel.getById(functionId)
+    })
+    .then(func => {
+      if (func._account !== accountId) throw errors.forbidden()
+      return func
     })
     .then(responses.ok)
     .catch(responses.error)
@@ -68,7 +100,7 @@ module.exports.populate = (event, context, callback) => {
         'eu-west-3',
         'sa-east-1' ]
 
-      let promises = []
+      let functions = []
       for (let i = 0; i < functionsCount; i++) {
         let func = {}
         let name = faker.hacker.verb() + ' ' + faker.hacker.noun()
@@ -80,14 +112,14 @@ module.exports.populate = (event, context, callback) => {
         func.name = name
         func.timeout = faker.random.number({min: 0, max: 300})
         let funcModel = new FunctionModel(func)
-        promises.push(funcModel.save())
+        functions.push(funcModel.data)
       }
-      return Promise.all(promises)
+      return functions
     })
-    .then(value => {
+    .then(functions => {
       let eventCounter = new BigInteger('33773592922032033623252972461104249685054501853208182784').plus(faker.random.number({min: 1000, max: 5000}))
-      let promisesInv = []
-      for (let func of value) {
+      let invocations = []
+      for (let func of functions) {
         let invocationsCount = faker.random.number({min: Math.min(invocationsMin, invocationsMax), max: Math.max(invocationsMax, invocationsMin)})
         let errorsCount = Math.round(errorsCoef * invocationsCount)
         for (let i = 0; i < invocationsCount; i++) {
@@ -104,6 +136,7 @@ module.exports.populate = (event, context, callback) => {
             '_id': _id,
             'billedDuration': billedDuration,
             'duration': duration,
+            'startTime': randomDateMillis,
             'logs': [
               {
                 'eventId': (eventCounter = eventCounter.plus(faker.random.number({min: 1, max: 1000}))).toString(),
@@ -171,14 +204,16 @@ module.exports.populate = (event, context, callback) => {
             errorsCount--
           }
           let invocationModel = new InvocationModel(invocation)
-          promisesInv.push(invocationModel.save())
+          invocations.push(invocationModel.data)
         }
       }
-      return Promise.all(promisesInv)
+      return {
+        functions: functions,
+        invocations: invocations
+      }
     })
-    .then(value => {
- //     console.log(value)
-      return undefined
+    .then(({functions, invocations}) => {
+      return Promise.all([FunctionModel.batchWrite(functions), InvocationModel.batchWrite(invocations)])
     })
     .then(responses.created)
     .catch(responses.error)
@@ -188,10 +223,12 @@ module.exports.populate = (event, context, callback) => {
 module.exports.clear = (event, context, callback) => {
   return Promise.all([InvocationModel.getAllScan({})
     .then(({Items}) => {
-      return InvocationModel.bulkDelete({Keys: Items.map((item) => { return {_id: item._id} })})
+      if (Items && Items.length > 0) { return InvocationModel.bulkDelete({Keys: Items.map((item) => { return {_id: item._id} })}) }
+      return []
     }), FunctionModel.getAllScan({})
     .then(({Items}) => {
-      return FunctionModel.bulkDelete({Keys: Items.map((item) => { return {_id: item._id} })})
+      if (Items && Items.length > 0) { return FunctionModel.bulkDelete({Keys: Items.map((item) => { return {_id: item._id} })}) }
+      return []
     })
   ])
     .then(responses.ok)
